+++ .dir-locals.el
@@ -1,1 +1,2 @@
-((nil . ((cider-clojure-cli-global-options . "-A:dev:test"))))
+((nil . ((cider-clojure-cli-global-options . "-A:dev:test")
+         (cider-default-cljs-repl          . node))))