+++ src/lambdaisland/glogi.cljs
@@ -45,13 +45,20 @@
   [lvl]
   (.-value (level lvl)))
 
+(defn make-log-record ^LogRecord [level message name exception]
+  (let [record (LogRecord. level message name)]
+    (when exception
+      (.setException record exception))
+    record))
+
 (defn log
   "Output a log message to the given logger, optionally with an exception to be
   logged."
   ([name lvl message]
    (log name lvl message nil))
   ([name lvl message exception]
-   (glog/log (logger name) (level lvl) message exception)))
+   (.logRecord (logger name)
+               (make-log-record (level lvl) message name exception))))
 
 (defn set-level
   "Set the level (a keyword) of the given logger, identified by name."